refactor(puzzle-generation): split prompt builder + response shape + composer
